requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
annotated-types==0.7.0
attrs==24.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
greenlet==3.0.1
idna==3.8
iniconfig==2.0.0
packaging==24.1
page==0.2
playwright==1.40.0
pluggy==1.5.0
pydantic==2.8.2
pydantic_core==2.20.1
pyee==11.0.1
pytest==7.4.3
pytest-base-url==2.1.0
pytest-playwright==0.5.1
python-slugify==8.0.4
requests==2.32.3
text-unidecode==1.3
typing_extensions==4.12.2
urllib3==2.2.2